pyproject.toml
[tool.poetry]
name = "repl_python3_FeelBetterApp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
ttkthemes = "^3.1.1"
pyttk = "^0.3.2"
googlemaps = "^4.4.2"
geocoder = "^1.38.1"
ipinfo = "^3.0.0"
praw = "^7.1.0"
pafy = "^0.5.5"
python-vlc = "^3.0.11115"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
